Group validation added

+++ app/controllers/admin/groups_controller.rb
@@ -36,17 +36,19 @@ class Admin::GroupsController < Admin::ApplicationController
   # POST /groups
   #----------------------------------------------------------------------------
   def create
-    # @group.attributes = group_params
-    # @group.save
     @group = Group.new
     @group.name = params[:group][:name]
     @group.company_id = current_user.company_id
 
-    if @group.save
-      @users_selected = params[:group][:user_ids].reject { |c| c.empty? } if params[:group][:user_ids].present?
-      @users_selected.uniq.each do |user_id|
-        @group.user_groups.create(user: User.where(id: user_id).first)
+    if params[:group][:user_ids].count > 1
+      if @group.save
+        @users_selected = params[:group][:user_ids].reject { |c| c.empty? } if params[:group][:user_ids].present?
+        @users_selected.uniq.each do |user_id|
+          @group.user_groups.create(user: User.where(id: user_id).first)
+        end
+        respond_with(@group)
       end
+    else
       respond_with(@group)
     end
   end
@@ -54,19 +56,22 @@ class Admin::GroupsController < Admin::ApplicationController
   # PUT /groups/1
   #----------------------------------------------------------------------------
   def update
-    # @group.update_attributes(group_params)
-    if @group.update_attributes(name: params[:group][:name])
-      @users_selected = params[:group][:user_ids].reject { |c| c.empty? } if params[:group][:user_ids].present?
-      @group.user_groups.each do |user_group|
-        unless @users_selected.include?(user_group.user_id.to_s)
-          user_group.destroy
+    if params[:group][:user_ids].count > 1
+      if @group.update_attributes(name: params[:group][:name])
+        @users_selected = params[:group][:user_ids].reject { |c| c.empty? } if params[:group][:user_ids].present?
+        @group.user_groups.each do |user_group|
+          unless @users_selected.include?(user_group.user_id.to_s)
+            user_group.destroy
+          end
         end
-      end
-      @users_selected.uniq.each do |user_id|
-        unless @group.user_groups.exists?(user_id: user_id)
-          @group.user_groups.create(user: User.where(id: user_id).first)
+        @users_selected.uniq.each do |user_id|
+          unless @group.user_groups.exists?(user_id: user_id)
+            @group.user_groups.create(user: User.where(id: user_id).first)
+          end
         end
+        respond_with(@group)
       end
+    else
       respond_with(@group)
     end
   end
